Move feed document to search app
