+++ src/feed/views/feed_view_v2.py
@@ -3,7 +3,6 @@ from opensearchpy import Q
 from feed.document_serializers.feed_document_serializer import (
     FeedEntryDocumentSerializer,
 )
-from feed.documents.feed_document import FeedEntryDocument
 from feed.views.common import FeedPagination
 from feed.views.feed_view_mixin import FeedViewMixin
 from search.base.filters import (
@@ -13,6 +12,7 @@ from search.base.filters import (
     SearchFilterBackend,
 )
 from search.base.viewsets import ElasticsearchViewSet
+from search.documents.feed import FeedEntryDocument
 
 
 class FeedV2ViewSet(FeedViewMixin, ElasticsearchViewSet):
